favorites/cleanup
* Encapsulated the Product.java class

* Created the foundation for filesaving

* Revert "Encapsulated the Product.java class"

This reverts commit 7529cec2

* Reverted back from encapsulating the Product class

* added test file to Fevorites

* Worked more on the file saving system

* Added some tests to the SaveableProductLists

* Added Methods and javadoc to ListIOManager

Added the apointed methods and javadoc to the class that is stated in the UML design diagram

* Changed the name of the Classes

* Changed return values

Changed the return value of load all, since it is supposed to return a list of SaveableProductLists

* NameChange of classes

Removed the classes with the old names

* Added Methods and javadoc

Altered and added some methods to the SaveableProductList. Added javadoc to all of the methods

* Added comment structure

* Added close method to the interface

Added a close method that is called by the manager, the close method the closes the buffered reader and writer

* The save method now writes to a file

The save method now writes the SavedProductList to the file

* Executes the close method to the shut down function in the IOManager

The program now calls the close method when the program shuts down

* getName for SaveableProductList

SaveableProductList now has a getter for the name

* Test list save function

Added a test to see if the program saves a list to the txt file

* The file that the lists are saved to

* The file handler can now read a list from a file

The file handler can nom read a list from a file, howevere i needed to change it so that the SaveableProducvtList now holds ids of products instead of products objects. Thi however, can be changed in the future when the filter is implemented

* Removed some TODOs

* Fixed some warnings

* Implemented the appropriate methods for the manager

Created and finalised the methods for the manager.

* Fixed som faulty comments

* Fixed some JavaDoc

* Added linking to javadoc

* Fixed some Javadoc

* Structured the tests

* Added Tests to SavableListClass

The test for SavableProductLits now cover 100%

* Worked on tests for favorites

* Started working on the equal condition for the SavableProductList

* Fixed an Equal method for SavableProductList

* Tests and bug fixes fo theProductListManager

* Connected the Datahandler to the FileIOManager

Added all the nessesary methods for the datahandler to use the file managment IO system

* Added test template for the new methods in Datahandler

* Fixed RemoveFromFavorites in datahandler

Added som todos as well for the Datahandler

* fixed scope issue by adding getters in Product

* resolved issue with previous merge conflict

* cleanup, comment out broken tests

* more broken tests

* reformatting

* Fixed all tests for Favorites

Co-authored-by: DanielRygaard <daniel.rygaard@live.se>
Co-authored-by: DanielRygaard <72791196+Danryg@users.noreply.github.com>
Co-authored-by: emilsvennesson <emilsvennesson@users.noreply.github.com>